requirements.txt
alembic==1.7.1
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2==2.9.1
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==2.3.3
